+++ inventory_components.tf
@@ -25,7 +25,8 @@ locals {
   instances_aws_transform = {
     for instance in var.instances_aws : lookup(instance.tags, "Name", instance.id) => merge(
       { "ansible_host" = instance.private_ip },
-      instance.tags
+      instance.tags,
+      can(instance.password_data) ? { "ansible_transport" = "winrm" } : {}
     )
   }
 
@@ -62,7 +63,8 @@ locals {
     for instance in var.instances_azr : instance.name => merge(
       { "ansible_host" = instance.private_ip_address },
       { "ansible_become_user" = try(instance.admin_ssh_key.0.username, instance.admin_username) },
-      instance.tags
+      instance.tags,
+      can(regex("Windows", instance.source_image_reference.0.offer)) ? { "ansible_transport" = "winrm" } : {}
     )
   }
 
@@ -81,7 +83,8 @@ locals {
     for instance in var.instances_vsp : instance.name => merge(
       { "ansible_host" = instance.default_ip_address },
       try({ "ansible_become_user" = instance.clone.0.customize.0.windows_options.0.full_name }, {}),
-      try(instance.vapp[0].properties, {})
+      try(instance.vapp[0].properties, {}),
+      can(instance.clone.0.customize.0.windows_options) ? { "ansible_transport" = "winrm" } : {}
     )
   }