+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-### 1.1.1
+### 1.1.1 (Next)
 - Add `inv_files` output.
 - Add `inv_file_perms` variable.
+- Attempt to detect if Windows instance, and set transport to `winrm` if true.
 
 ### 1.1.0
 - Update `instances` var to `map(set(object))` to support Ansible group specification.